Optimised the headers removing used and unnecessary ones. The compile time should be much better now.
Changed some default values so they better represent the underlying type.
