+++ src/hades_extensions/include/generation/map.hpp
@@ -3,11 +3,12 @@
 
 // Std headers
 #include <optional>
+#include <random>
+#include <unordered_map>
 #include <unordered_set>
 
 // Local headers
-#include "bsp.hpp"
-#include "searching.hpp"
+#include "primitives.hpp"
 
 // ----- STRUCTURES ------------------------------
 /// Represents an undirected weighted edge in a graph.
