--- src/hades_extensions/include/hash_combine.hpp
@@ -1,18 +1,0 @@
-// Ensure this file is only included once
-#pragma once
-
-// Std headers
-#include <cstddef>
-#include <functional>
-
-// ----- FUNCTIONS ------------------------------
-/// Allows multiple hashes to be combined for a struct
-///
-/// @tparam T - The type of the value to hash.
-/// @param seed - The seed for initialising the hasher.
-/// @param value - The value to hash.
-template <typename T>
-void hash_combine(std::size_t &seed, const T &value) {
-  std::hash<T> hasher;
-  seed ^= hasher(value) + 0x9e3779b9 + (seed << 6) + (seed >> 2);  // NOLINT
-}
