+++ src/hades_extensions/src/game_objects/systems/armour_regen.cpp
@@ -2,6 +2,7 @@
 #include "game_objects/systems/armour_regen.hpp"
 
 // Local headers
+#include "game_objects/registry.hpp"
 #include "game_objects/stats.hpp"
 
 // ----- CONSTANTS -------------------------------
