+++ src/hades_extensions/src/game_objects/systems/movements.cpp
@@ -8,6 +8,7 @@
 #include <chipmunk/chipmunk_structs.h>
 
 // Local headers
+#include "game_objects/registry.hpp"
 #include "game_objects/systems/physics.hpp"
 
 // ----- CONSTANTS ------------------------------
