+++ src/hades_extensions/tests/game_objects/systems/test_inventory.cpp
@@ -1,4 +1,5 @@
 // Local headers
+#include "game_objects/registry.hpp"
 #include "game_objects/systems/inventory.hpp"
 #include "macros.hpp"
