+++ src/hades_extensions/tests/game_objects/systems/test_movements.cpp
@@ -5,7 +5,6 @@
 #include <numbers>
 
 // Local headers
-#include "game_objects/stats.hpp"
 #include "game_objects/systems/movements.hpp"
 #include "game_objects/systems/physics.hpp"
 #include "macros.hpp"
